The records view look good.

+++ NeverSkipLegDay/NeverSkipLegDay/NeverSkipLegDay/Models/Record.cs
@@ -8,6 +8,6 @@ namespace NeverSkipLegDay.Models
         public int Id { get; set; }
         public int ExerciseId { get; set; }
         public int? Reps { get; set; }
-        public int? Weight { get; set; }
+        public decimal? Weight { get; set; }
     }
 }
